Enhance NotAuthorized page with user info and simplify JS
The changes in `NotAuthorized.cshtml` introduce the use of
`Microsoft.AspNetCore.Identity` and `Microsoft.Extensions.DependencyInjection`
namespaces to access user information through the `UserManager` service.
`ViewData["Title"]` is set to "NotAuthorized". If the user is authenticated,
the code retrieves the user's details and sets `ViewBag.ImageFullPath` and
`ViewBag.FirstName` with the user's profile image path and first name,
respectively. The HTML structure is updated to conditionally display the
user's profile picture or first name, or a default user div if the user is
not authenticated.

In `notauthorized.js`, the call to add the 'scanning' class to the
`.scan-window` element is removed from the `tl.call` function, simplifying
the scanning animation.