Refactor AccountController methods to be asynchronous
Refactored the `Login` and `Register` methods in the `AccountController` class to be asynchronous. Simplified user authentication checks by removing the `this` keyword. Updated the logic to log out authenticated users asynchronously using `_userHelper.LogoutAsync()` and redirect them to the appropriate actions ("Login" for `Login` method and "Register" for `Register` method) instead of the "Index" action of the "Home" controller.